Календарь: четыре технических сезона в механике
Этап 1а плана живой растительности. Год 365 суток от 1 марта, сезоны по
месяцам как у UDS, по лору осень -- Лето. Летние травы -- июнь–август,
осенние и Листовики -- вся осень, Купальская ночь -- на 24 июня, осень
нейтральна. Настройки окон внутри Лета удалены. Игровые часы -- double:
во float они вставали в конце ноября. GDD §15.4 переписан, замер §15.8
пересчитан. Тесты 655/0 дважды.

+++ ProjectHerbalist/Source/ProjectHerbalist/Core/Save/HerbalistSaveTypes.h
@@ -224,9 +224,11 @@ public:
     // Игровые часы (AGridWorldManager::GameClockSeconds) — фаза суток и
     // будущая погода (UltraDynamicSky, ROADMAP.md "Реальный Ultra Dynamic
     // Weather") должны возобновляться с той же точки, а не с рассвета
-    // каждой новой сессии.
+    // каждой новой сессии. double с 2026-09-16 (см. AGridWorldManager::
+    // GetGameClockSeconds); старые сейвы с float читаются тегированной
+    // сериализацией без потерь, версия не поднималась.
     UPROPERTY()
-    float GameClockSeconds = 0.0f;
+    double GameClockSeconds = 0.0;
 
     UPROPERTY()
     TArray<FSavedCellState> Cells;
